Add then function
